Move Jetpack constants to own file
